% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{methods}
\name{mgroup}
\alias{mgroup}
\alias{mgroup,MongoPipeline-method}
\title{Add a $group stage to a MongoPipeline.}
\usage{
mgroup(p, by, ...)

\S4method{mgroup}{MongoPipeline}(p, by, ...)
}
\arguments{
\item{p}{A \linkS4class{MongoPipeline} instance.}

\item{by}{The grouping variable(s), as an expression.}

\item{...}{Zero or more name=value pairs where name represent
a field to create, and values are accumulator expressions.}
}
\value{
A copy of the previous pipeline, with the new stage added.
}
\description{
Add a $group stage to a MongoPipeline.
}
