% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{methods}
\name{mmatch}
\alias{mmatch}
\alias{mmatch,MongoPipeline-method}
\title{Add a $match stage to a MongoPipeline.}
\usage{
mmatch(p, expr)

\S4method{mmatch}{MongoPipeline}(p, expr)
}
\arguments{
\item{p}{A \linkS4class{MongoPipeline} instance.}

\item{expr}{An expression indicating the requirements to match against.}
}
\value{
A copy of the previous pipeline, with the new stage added.
}
\description{
Add a $match stage to a MongoPipeline.
}
