% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoplyr-package.R
\docType{package}
\name{mongoplyr-package}
\alias{mongoplyr}
\alias{mongoplyr-package}
\title{mongoplyr: Use {dplyr} and {dbplyr} with MongoDB}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use {dplyr} verbs to query a MongoDB database. This uses {dbplyr} to create SQL queries, and then converts them into MongoDB JSON queries with "mongotranslate" from the MongoDB BI connector (to be installed). One can also recover the JSON query to use it directly into {mongolite}. This way, {mongoplyr} serves as a translator from "dplyr" code to MongoDB JSON queries.
}
\details{
Use {dplyr} verbs to query a MongoDB database. This uses {dbplyr} to create
SQL queries, and then converts them into MongoDB JSON queries with
"mongotranslate" from the MongoDB BI Connector (to be installed; Make sure
you have adequate license to use it).

One can also recover the JSON query to use it directly into "mongolite". This
way, {mongoplyr} serves as a translator from "dplyr" code to MongoDB JSON
queries. One the JSON string is constructed, use it with {mongolite} and you
do not need {mongoplyr} any more to run your code.

Before you can use this package, you must install "mongotranslate" and
"mongodrdl" from MongoDB and make them available on your PATH (or specify the
directory where they are in \code{options(mongotranslate.path = ....)}). Follow
instructions to install BI Connector on Premise on your system. Note that
"mongotranslate" is apparently not available for Windows. Consequently, the
{mongoplyr} package is useless on this OS. Use Linux or MacOS (use WLS on
Windows, or a Docker container or a virtual machine instead).
}
\section{Important functions}{

\itemize{
\item \code{\link[=tbl_mongo]{tbl_mongo()}} creates an object that connects to your MongoDB database and
use {dplyr} verbs in a lazy way.
\item \code{\link[=time_from_id]{time_from_id()}} get the creation date of a MongoDB document from its
ObjectID.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/mongoplyr}
  \item \url{https://www.sciviews.org/mongoplyr/}
  \item Report bugs at \url{https://github.com/SciViews/mongoplyr/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
