% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{methods}
\name{mproject}
\alias{mproject}
\alias{mproject,MongoPipeline-method}
\title{Add a $project stage to a MongoPipeline.}
\usage{
mproject(p, ...)

\S4method{mproject}{MongoPipeline}(p, ...)
}
\arguments{
\item{p}{A \linkS4class{MongoPipeline} instance.}

\item{...}{Named arguments for mproject fields, indicating whether to include (1) or exclude (0)}
}
\value{
A copy of the previous pipeline, with the new stage added.
}
\description{
Add a $project stage to a MongoPipeline.
}
