% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{methods}
\name{msort}
\alias{msort}
\alias{msort,MongoPipeline-method}
\title{Add a $sort stage to a MongoPipeline.}
\usage{
msort(p, ...)

\S4method{msort}{MongoPipeline}(p, ...)
}
\arguments{
\item{p}{A \linkS4class{MongoPipeline} instance.}

\item{...}{Named arguments for sort fields, indicating whether to sort in ascending (1) or descending (-1) order.}
}
\value{
A copy of the previous pipeline, with the new stage added.
}
\description{
Add a $sort stage to a MongoPipeline.
}
