% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{methods}
\name{munwind}
\alias{munwind}
\alias{munwind,MongoPipeline-method}
\title{Add an $unwind stage to a MongoPipeline.}
\usage{
munwind(p, expr)

\S4method{munwind}{MongoPipeline}(p, expr)
}
\arguments{
\item{p}{A \linkS4class{MongoPipeline} instance.}

\item{expr}{An expression indicating the array field to unwind.}
}
\value{
A copy of the previous pipeline, with the new stage added.
}
\description{
Add an $unwind stage to a MongoPipeline.
}
