% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_from_id.R
\name{time_from_id}
\alias{time_from_id}
\title{Get timestamp for ObjectID of MongoDB documents}
\usage{
time_from_id(id, tz = "GMT")
}
\arguments{
\item{id}{A vector with one or more objid.}

\item{tz}{The timezone to use (GMT by default).}
}
\value{
A **POSIXct object with time extracted from the ObjectID.
}
\description{
Each MongoDB document has an ObjectID. When it is automatically
constructed (it is almost always the case), the first part of the ObjectID
is a timestamp indicating when the document was added. \code{\link[=time_from_id]{time_from_id()}}
recovers this timestamp.
}
\examples{
# ObjectID 5a682326bf8380e6e6584ba5 was created at
# ISODate("2018-01-24T06:09:42Z")
time_from_id("5a682326bf8380e6e6584ba5")
# Get it in West Europe time
time_from_id("5a682326bf8380e6e6584ba5", tz = "CET")
}
