% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{methods}
\name{mexecute}
\alias{mexecute}
\alias{mexecute,MongoPipeline-method}
\title{Execute a MongoPipeline's current query.}
\usage{
mexecute(p, client)

\S4method{mexecute}{MongoPipeline}(p, client)
}
\arguments{
\item{p}{A \linkS4class{MongoPipeline} instance.}

\item{client}{A connection to the mongo collection to execute against.}
}
\value{
A data frame containing the aggregation pipeline results.
}
\description{
Execute a MongoPipeline's current query.
}
