% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo.R
\docType{class}
\name{MongoPipeline-class}
\alias{MongoPipeline-class}
\alias{MongoPipeline}
\title{Represents a MongoDB aggregation pipeline query.}
\description{
Represents a MongoDB aggregation pipeline query.
}
\section{Slots}{

\describe{
\item{\code{stages}}{Character vector containing the current pipeline stages, serialised as JSON.}
}}

