% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_mongo.R
\name{tbl_mongo}
\alias{tbl_mongo}
\alias{collapse.tbl_mongo}
\alias{collect.tbl_mongo}
\alias{print.mongo_query}
\title{Create a tbl_mongo lazy connection to a MongoDB collection}
\usage{
tbl_mongo(
  collection = "test",
  db = "test",
  url = "mongodb://localhost",
  max_scan = 1000L,
  ...,
  mongotranslate.path = getOption("mongotranslate.path")
)

\method{collapse}{tbl_mongo}(x, keep.names = FALSE, ...)

\method{collect}{tbl_mongo}(x, keep.names = FALSE, ...)

\method{print}{mongo_query}(x, sql = FALSE, ...)
}
\arguments{
\item{collection}{The collection to use in the MongoDB database.}

\item{db}{The database to use from the MongoDB server.}

\item{url}{The URL to the MongoDB server. This uses \code{\link[=mongo]{mongo()}} from
{mongolite} internally, see the documentation at
https://jeroen.github.io/mongolite/connecting-to-mongodb.html.}

\item{max_scan}{The maximum number of documents to scan in the collection in
order to infer the corresponding schema with mongodrdl (1000 by default).}

\item{...}{More parameters to \code{\link[=mongo]{mongo()}} to connect to the MongoDB server.}

\item{mongotranslate.path}{The path to the mongotranslate and mongodrdl
software. Can be set via \code{options(mongotranslate.path = ....)}, or left empty
if these executables are on the search path.}

\item{x}{A \strong{mongo_query} object as obtained with \code{\link[=collapse]{collapse()}}.}

\item{keep.names}{Logical (\code{FALSE} by default). Should the (strange) names
constructed by {dbplyr} be kept in the JSON MongoDB query or not?}

\item{sql}{Should the corresponding SQL statement be printed as well as the
JSON query (\code{FALSE} by default?}
}
\value{
A \strong{tbl_mongo} object that contains the logic to process queries on
a MongoDB collection through {dplyr} verbs. \code{\link[=collect]{collect()}} returns a data.frame
with the result from querying the MongoDB collection. \code{\link[=collapse]{collapse()}} returns
the MongoDB JSON query corresponding to the process in a \strong{mongo_query}
object.
}
\description{
The \strong{tbl_mongo} object is a lazy connection to a MongoDB
collection that you can use with {dbplyr} and {dplyr} verbs. The query
in MongoDB JSON language is computed on \code{\link[=collect]{collect()}}ing the results, or
by using \code{\link[=collapse]{collapse()}} to retrieve the JSON command in a character string.
}
\examples{
# TODO...
}
